refactor: extract AccumulatorContinue companion methods to extension functions
AccumulatorContinue was hidden inside Modular.kt in org.elixir_lang.psi - a general-purpose iteration primitive buried in a domain-specific file and package.

This change:
- Moves AccumulatorContinue to org.elixir_lang.util where infrastructure belongs
- Strips it to a pure data class with no companion object
- Extracts foldWhile as idiomatic Kotlin extension functions on Iterable, Sequence, and Array in their own FoldWhile.kt file
- Moves childExpressionsFoldWhile to a PsiElement extension in psi.impl where the PSI dependency belongs

Call sites now read as `collection.foldWhile(initial) { ... }` rather than `AccumulatorContinue.foldWhile(collection, initial) { ... }`, matching stdlib conventions like fold, map, and filter.

Also remove dead `org.elixir_lang.beam.psi.Modular` interface
