--- src/org/elixir_lang/beam/psi/Modular.java
@@ -1,4 +1,0 @@
-package org.elixir_lang.beam.psi;
-
-public interface Modular {
-}
