+++ src/org/elixir_lang/util/AccumulatorContinue.kt
@@ -1,0 +1,3 @@
+package org.elixir_lang.util
+
+data class AccumulatorContinue<out R>(val accumulator: R, val `continue`: Boolean)
